fixed seed file, setting up redux connection for cart component
