+++ client/redux/action-types/cart.js
@@ -1,0 +1,1 @@
+export const GET_CART_PRODUCTS = 'GET_CART_PRODUCTS'
