+++ seed.js
@@ -1,7 +1,7 @@
-const Users = require('./server/db/models/user')
-const Products = require('./server/db/models/product')
-const Orders = require('./server/db/models/order')
-const db = require('./server/db/db')
+'use strict'
+
+const db = require('./server/db')
+const {User, Product, Order} = require('./server/db/models')
 const faker = require('faker')
 
 const instanceCount = 100
@@ -54,36 +54,48 @@ const dummyCandies = [
   }
 ]
 
-const seed = async () => {
+async function seed() {
+  await db.sync({force: true})
+  console.log('db synced!')
+
+  await Promise.all(
+    users.map(user => {
+      return User.create(user)
+    })
+  )
+
+  await Promise.all(
+    dummyCandies.map(candy => {
+      return Product.create(candy)
+    })
+  )
+
+  console.log(`seeded successfully`)
+}
+
+// We've separated the `seed` function from the `runSeed` function.
+// This way we can isolate the error handling and exit trapping.
+// The `seed` function is concerned only with modifying the database.
+async function runSeed() {
+  console.log('seeding...')
   try {
-    await db.sync({force: true})
-
-    await Promise.all(
-      users.map(user => {
-        return Users.create(user)
-      })
-    )
-
-    await Promise.all(
-      dummyCandies.map(candy => {
-        return Products.create(candy)
-      })
-    )
+    await seed()
   } catch (err) {
     console.error(err)
+    process.exitCode = 1
+  } finally {
+    console.log('closing db connection')
+    await db.close()
+    console.log('db connection closed')
   }
 }
 
-module.exports = seed
-
-if (require.main === module) {
-  seed()
-    .then(() => {
-      console.log('Seeding success!')
-      db.close()
-    })
-    .catch(err => {
-      console.error(err)
-      db.close()
-    })
+// Execute the `seed` function, IF we ran this module directly (`node seed`).
+// `Async` functions always return a promise, so we can use `catch` to handle
+// any errors that might occur inside of `seed`.
+if (module === require.main) {
+  runSeed()
 }
+
+// we export the seed function for testing purposes (see `./seed.spec.js`)
+module.exports = seed
