+++ server/db/models/index.js
@@ -3,11 +3,11 @@ const Product = require('./product')
 const Order = require('./order')
 const Cart = require('./cart')
 
+User.hasMany(Order)
+Order.belongsTo(User, {as: 'user'})
+
 Product.belongsToMany(Order, {through: Cart})
 Order.belongsToMany(Product, {through: Cart})
-
-User.hasMany(Order)
-Order.belongsTo(User)
 
 module.exports = {
   User,
